fastapi-jwt-ai/requirements.txt
fastapi
uvicorn
python-jose[cryptography]
pydantic
requests
pytest
